aider: Added import statement for `debug` macro in integrity zome scope.
# Aider chat conversation:

USER: cannot find macro debug in integrity zome scope

ASSISTANT: To resolve the issue of the `debug` macro not being found in the integrity zome's scope, we need to ensure that the `hdk::prelude` module is imported, which includes the `debug!` macro. If it's not already imported, we will add the necessary import statement. If it is imported, we will check for any conditional compilation flags or other issues that might be preventing the macro from being recognized.

Let's add the necessary import statement to the `generator.rs` file:

```diff
--- happ/dnas/asset_validator/zomes/integrity/validation_claims/src/generator.rs
+++ happ/dnas/asset_validator/zomes/integrity/validation_claims/src/generator.rs
@@ ... @@
 use hdi::prelude::*;

+use hdk::prelude::debug;
+
 #[hdk_entry_helper]
 #[derive(Clone, PartialEq)]
 pub struct Generator {
```

This change will make the `debug!` macro available in the scope of the integrity zome, allowing it to be used within the `validate_update_generator` function and elsewhere in the zome.