aider: Updated ethAddress assignment to occur only on button click.
# Aider chat conversation:

USER: change it so the ethAddress doesn't get updated until the button is clicked

ASSISTANT: To change the behavior so that the `ethAddress` doesn't get updated until the button is clicked, I will modify the input field to use a local variable that is only assigned to `ethAddress` when the register button is clicked. Here are the diffs for the changes:

```diff
--- happ/ui-user/src/LoginOrRegister.svelte
+++ happ/ui-user/src/LoginOrRegister.svelte
@@ ... @@
   let ethAddress = '';
+  let tempEthAddress = '';
   let keyPair;
   let signingKey;
   let errorMessage = '';
   let qrCodeImage = '';
   let tpDeepLink = '';
@@ ... @@
   async function register() {
+    ethAddress = tempEthAddress;
     try {
         let payload = {
           handle,
           signingKey: encodeHashToBase64(signingKey),
         }
@@ ... @@
         <div>
             If token proof isn't working submit this form with an email, ENS, or eth address. Just remember it will be public in the hApp.
-            <input type="text" value={ethAddress} on:input={e => ethAddress = e.target.value} placeholder="ethAddress, email, ENS" required>
+            <input type="text" bind:value={tempEthAddress} placeholder="ethAddress, email, ENS" required>
             <button on:click={register}>Register</button>
         </div>
   {/if}
{/if}
```

This change introduces a new variable `tempEthAddress` that is bound to the input field. The `ethAddress` is only updated with the value of `tempEthAddress` when the `register` function is called, which happens when the register button is clicked.